requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
idna==3.6
lxml==5.1.0
magic-filter==1.0.12
multidict==6.0.5
outcome==1.3.0.post0
pydantic==2.5.3
pydantic_core==2.14.6
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.25.0
trio-websocket==0.11.1
typing==3.7.4.3
typing_extensions==4.10.0
undetected-chromedriver==3.5.5
urllib3==2.2.1
websockets==12.0
wsproto==1.2.0
yarl==1.9.4
